Add order button and change margin to top section
Add a order now button to the top section and the change margin

I forgot there was a button in the hero section so I added one in. I
also did not like the way things looked in the hero section so I changed
the bottom margin which made things look better to me.

 Changes to be committed:
	modified:   css/style.css
	modified:   index.html